Flatten hooks folder
