--- src/controllers/hooks/index.ts
@@ -1,8 +1,0 @@
-export { useResource } from './resource-store';
-export {
-  useRouter,
-  useRouterActions,
-  useQueryParam,
-  usePathParam,
-} from './router-store';
-export { useTimeout } from './use-timeout';
