+++ src/controllers/index.ts
@@ -1,18 +1,9 @@
-export {
-  useResource,
-  useRouter,
-  useRouterActions,
-  useQueryParam,
-  usePathParam,
-  useTimeout,
-} from './hooks';
-
-export { Router } from './router';
-export { RouterActions } from './router-actions';
-export { RouterSubscriber } from './router-subscriber';
 export { MemoryRouter } from './memory-router';
 export { Redirect } from './redirect';
 export { ResourceSubscriber } from './resource-subscriber';
+export { Router } from './router';
+export { RouterActions } from './router-actions';
+export { RouterSubscriber } from './router-subscriber';
 export { StaticRouter } from './static-router';
 
 export {
@@ -20,6 +11,7 @@ export {
   useResourceStoreContext,
   ResourceDependencyError,
 } from './resource-store';
+
 export type {
   CreateResourceArgBase,
   CreateResourceArgSync,
@@ -30,5 +22,11 @@ export {
   RouteResourceEnabledSubscriber,
   createRouterSelector,
 } from './router-store';
+
+export { usePathParam } from './use-path-param';
+export { useQueryParam } from './use-query-param';
+export { useResource } from './use-resource';
+export { useRouter } from './use-router';
+export { useRouterActions } from './use-router-actions';
 
 export { withRouter } from './with-router';
